+++ client/projects/cms/src/app/modules/dashboard/modules/module-instance/components/export/export.component.ts
@@ -102,8 +102,6 @@ export class ExportComponent {
         }));
       }
 
-      console.log(222, this.data.collection);
-
       return from(
         firebase.auth().currentUser.getIdToken()
       )